app/requirements.txt
pandas
numpy
beautifulsoup4
sqlalchemy  # it is an orm 
python-dotenv
requests
uvicorn[standard]
fastapi
transformers
torch
matplotlib
httpx
feedparser
openai